Product CRUD and Details

+++ src/main/java/posojt/persistence/entity/Product.java
@@ -42,7 +42,7 @@ public class Product {
 	@Column(nullable = false)
 	private int count;
 	
-	private int size;
+	private float size;
 	
 	private String color;
